requirements.txt
absl-py==0.15.0
alembic==1.8.0
astor==0.8.1
astunparse==1.6.3
attrs==21.4.0
autopage==0.5.1
backcall==0.2.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
clang==5.0
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
colorlog==6.6.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
filelock==3.7.1
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
gviz-api==1.10.0
h5py==3.1.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.8.0
ipykernel==6.13.0
ipython==7.34.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.1
jupyter-core==4.10.0
keras==2.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
lightgbm==3.3.1
Mako==1.2.1
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.19.5
oauthlib==3.2.0
opt-einsum==3.3.0
optuna==2.10.1
packaging==21.3
pandas==1.3.4
parso==0.8.3
pbr==5.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prettytable==3.3.0
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.8
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.0.0
regex==2022.7.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.4.1
seaborn==0.11.2
six==1.15.0
SQLAlchemy==1.4.39
stevedore==3.5.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-profile==2.8.0
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.16.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.1
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tokenizers==0.12.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
transformers==4.20.1
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.7.0
